requirements.txt
aiogram==3.4.1
fastapi==0.109.2
uvicorn[standard]==0.27.1
aiosqlite==0.19.0
apscheduler==3.10.4
pydantic==2.5.3
python-dotenv==1.0.1